+++ src/Schema/TableSchemaInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Schema;
 
-use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
 
 /**
@@ -202,17 +201,4 @@ interface TableSchemaInterface
      * @param array $to The foreign key.
      */
     public function foreignKey(string|int $id, array $to): void;
-
-    /**
-     * Set composite foreign key.
-     *
-     * @param int $id The index of foreign key.
-     * @param string $from The column name in current table.
-     * @param string $to The column name in foreign table.
-     *
-     * @throws NotSupportedException
-     *
-     * @deprecated will be removed in version 2.0.0
-     */
-    public function compositeForeignKey(int $id, string $from, string $to): void;
 }